コンポーネント分割の再調整および各ファイルの微調整完了
